fix: propagate duplex transform sink errors
Sink errors were either getting smothered or causing unhandled
promise rejections depending on your platform.

Instead if the sink of a duplex stream being used as a transform
returns a promise, propagate errors thrown by the sink into it's
source so they can be caught and handled by the consumer.